refactor: seperate classes
