+++ src/main/kotlin/blackjack/Main.kt
@@ -1,0 +1,3 @@
+package blackjack
+
+class Main
