+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,0 +1,6 @@
+package blackjack.domain
+
+class Card(
+    val rank: Rank,
+    val suit: Suit,
+)
