+++ src/main/kotlin/blackjack/domain/Rank.kt
@@ -1,0 +1,5 @@
+package blackjack.domain
+
+interface Rank {
+    val possibleValues: List<Int>
+}
