+++ src/main/kotlin/blackjack/domain/Result.kt
@@ -1,0 +1,8 @@
+package blackjack.domain
+
+enum class Result {
+    WIN,
+    DRAW,
+    LOSE,
+    NOT_YET,
+}
