+++ src/test/kotlin/blackjack/domain/NumberTest.kt
@@ -1,24 +1,8 @@
-package blackjack.test
+package blackjack.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-
-class Number(
-    value: Int,
-) : Rank {
-    override val possibleValues: List<Int> = listOf(value)
-
-    init {
-        require(possibleValues.all { possibleValue: Int -> possibleValue in RANGE })
-    }
-
-    companion object {
-        private const val MIN = 2
-        private const val MAX = 10
-        val RANGE = MIN..MAX
-    }
-}
 
 class NumberTest {
     @Test
