+++ src/test/kotlin/blackjack/domain/PlayerTest.kt
@@ -1,47 +1,8 @@
-package blackjack.test
+package blackjack.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-
-class Player(
-    val name: String,
-) {
-    private val hand: Hand = Hand(emptyList())
-    var wantToHit: Boolean? = null
-    var result: Result = Result.NOT_YET
-
-    fun getCard(card: Card) {
-        hand.add(card)
-    }
-
-    fun getCards(cards: List<Card>) {
-        hand.add(cards)
-    }
-
-    fun getCountOfCards(): Int = hand.getSize()
-
-    fun hitOrStay(hit: () -> Unit) {
-        if (wantToHit == true) {
-            hit()
-        }
-    }
-
-    fun setResult() {
-        if (hand.getScore() == null) {
-            result = Result.LOSE
-        }
-    }
-
-    fun getScore() = hand.getScore()
-}
-
-enum class Result {
-    WIN,
-    DRAW,
-    LOSE,
-    NOT_YET,
-}
 
 class PlayerTest {
     @Test
