+++ src/test/kotlin/blackjack/domain/RankTest.kt
@@ -1,11 +1,7 @@
-package blackjack.test
+package blackjack.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-
-interface Rank {
-    val possibleValues: List<Int>
-}
 
 class RankTest {
     @Test
